Fix update bug when loaded from new locale
When you load a pureted record with a different locale (e.g. `de`) than it was created (e.g. `en`), and you try to update it, it creates a new `{ :de => {} }` hash to its `puret_attributes`, because of the side-effectful init of puret_attributes (` @puret_attributes ||= Hash.new { |hash, key| hash[key] = {} }` :poop:

That causes the `#update_translations!` after_save hook to try and save that empty new de translation. If that translation has a `presence` validaiton for some fields, as our `ArticleTranslation` has in jobteaser, it just blows up.